+++ packages/expo-sms/android/src/test/java/expo/modules/sms/MockActivityProvider.kt
@@ -1,6 +1,0 @@
-package expo.modules.sms
-
-import expo.modules.core.interfaces.ActivityProvider
-import expo.modules.core.interfaces.InternalModule
-
-internal interface MockActivityProvider : ActivityProvider, InternalModule